refactor(interaction-create): ensure SRP
Every events should only have single responsibility.
